copy image

+++ sharp/zip/remove.js
@@ -2,6 +2,7 @@ const path = require('path')
 
 const fs = require('fs-extra')
 const { deepTraverse } = require('./deepTraverse')
+const { isImage } = require('../../app/utils/index')
 
 // 清理size === 0 损坏的图片
 async function remove(pathLike) {
@@ -11,7 +12,7 @@ async function remove(pathLike) {
     const file = allFiles[i]
 
     const { size } = fs.statSync(file)
-    if (size < 1) {
+    if (isImage(file) && size < 1) {
       result.push(file)
       await fs.remove(file)
     }
